requirements.txt
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
huggingface-hub==0.33.0
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-google-genai==2.1.5
langchain-openai==0.3.23
langchain-text-splitters==0.3.8
langsmith==0.3.45
llama_cpp_python==0.3.9
openai==1.86.0
pydantic
pydantic-settings
python-dotenv
PyYAML
regex
requests
SQLAlchemy